Use Python script for PHPDoc generation. #321
New PHP files were not being picked up automatically for PHPDoc
generation as they were specified explicitly in the Makefile. This
commit replaces the explicit specification of PHP files in the Makefile
with a Python script that searches the project for qualifying files and
invokes the doxphp command to generate the Sphinx RST output.